Fix wide char input. KeyChar should be unsigned for full 16-bit unicode

+++ src/common/engine/d_event.cpp
@@ -250,8 +250,8 @@ FUiEvent::FUiEvent(const event_t *ev)
 		IsCtrl = !!(ev->data3 & GKM_CTRL);
 		break;
 	case EV_GUI_Char:
-		KeyChar = ev->data1;
-		KeyString = MakeUTF8(ev->data1);
+		KeyChar = (uint16_t)ev->data1;
+		KeyString = MakeUTF8((uint16_t)ev->data1);
 		IsAlt = !!ev->data2; // only true for Win32, not sure about SDL
 		break;
 	default: // mouse event
